Refactored main app into own directory
