+++ app/common/startup.js
