+++ app/main/components/App.js
