+++ app/main/components/AuthenticationManager.js
