+++ app/main/components/LoginDialog.js
@@ -8,7 +8,7 @@ import RaisedButton from 'material-ui/RaisedButton'
 import TextField from 'material-ui/TextField'
 
 import { loginUser } from '../actions'
-import t from '../i18n'
+import t from '../../common/i18n'
 
 // TODO: Clear loginDialogError after successful login
 const reduxWrapper = connect(
